+++ include/structs.h
@@ -6,7 +6,7 @@
 /*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/21 08:37:38 by jkauker           #+#    #+#             */
-/*   Updated: 2024/06/29 20:02:27 by freddy           ###   ########.fr       */
+/*   Updated: 2024/06/29 21:16:40 by freddy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -91,6 +91,7 @@ typedef struct s_textures
 	mlx_texture_t	*door;
 	mlx_texture_t	*key;
 	mlx_texture_t	*goal;
+	mlx_texture_t	*health;
 
 	// Orbs
 	mlx_texture_t	*water_orb;
@@ -158,7 +159,7 @@ typedef struct s_player
 	t_transform	transform;
 	t_transform	spawn_transform;
 
-	int			elemental_type;
+	int			element;
 
 	int			health;
 
@@ -174,7 +175,8 @@ typedef enum e_entity_type
 	KEY_NTT,
 	ORB_NTT,
 	PROJECTILE_NTT,
-	BLIGHT_NTT
+	BLIGHT_NTT,
+	HEALTH_NTT
 }	t_entity_type;
 
 typedef struct s_entity	t_entity;
@@ -222,14 +224,14 @@ typedef struct s_door
 
 typedef struct s_orb
 {
-	int	type;
+	int	element;
 }	t_orb;
 
 // - PROJECTILE
 
 typedef struct s_projectile
 {
-	int	type;
+	int	element;
 }	t_projectile;
 
 // - BLIGHT
@@ -241,7 +243,7 @@ typedef struct s_projectile
 
 typedef struct s_blight
 {
-	int	type;
+	int	element;
 	int	state;
 }	t_blight;