+++ src/3_game_loop/a_game_logic/entities/player.c
@@ -6,7 +6,7 @@
 /*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/29 19:57:01 by freddy            #+#    #+#             */
-/*   Updated: 2024/06/29 20:11:30 by freddy           ###   ########.fr       */
+/*   Updated: 2024/06/29 21:01:55 by freddy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ void	on_collision_player(t_entity	*collider)
 
 	if (!(collider->type == PROJECTILE_NTT))
 		return ;
-	if (a_beats_b(((t_projectile *)collider->data)->type, player()->elemental_type))
+	if (a_beats_b(((t_projectile *)collider->data)->element, player()->element))
 		health_decrease_amount = MAJOR_PROJECTILE_HIT_DESCREASE;
 	else
 		health_decrease_amount = MINOR_PROJECTILE_HIT_DECREASE;